:recycle:  "아키텍처 구조 개선"
