+++ src/main/java/spring/fitlinkbe/infra/common/token/TokenRepositoryImpl.java
@@ -1,9 +1,10 @@
-package spring.fitlinkbe.infra.common;
+package spring.fitlinkbe.infra.common.token;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import spring.fitlinkbe.domain.common.TokenRepository;
 import spring.fitlinkbe.domain.common.model.Token;
+import spring.fitlinkbe.infra.common.PersonalDetailJpaRepository;
 import spring.fitlinkbe.infra.common.model.PersonalDetailEntity;
 import spring.fitlinkbe.infra.common.model.TokenEntity;
