+++ src/main/java/spring/fitlinkbe/support/security/provider/GoogleProviderHandler.java
@@ -1,4 +1,4 @@
-package spring.fitlinkbe.infra.security.provider;
+package spring.fitlinkbe.support.security.provider;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
@@ -7,34 +7,27 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import spring.fitlinkbe.domain.common.model.PersonalDetail;
 
-import java.util.Map;
-
 import static spring.fitlinkbe.domain.common.model.PersonalDetail.OauthProvider;
 
 @Component
 @Transactional
 @RequiredArgsConstructor
-public class KakaoProviderHandler implements Oauth2ProviderHandler {
+public class GoogleProviderHandler implements Oauth2ProviderHandler {
     @Override
     public boolean supports(OauthProvider provider) {
-        return provider == OauthProvider.KAKAO;
+        return provider == OauthProvider.GOOGLE;
     }
 
     @Override
     public PersonalDetail handle(OAuth2UserRequest userRequest, OAuth2User user) {
-        Map<String, String> properties = user.getAttribute("properties");
-        assert properties != null;
-
-        Long id = user.getAttribute("id");
-        assert id != null;
-
-        // todo: email 받을 수 있도록 kakao 비즈앱 신청 후 추가 (FitLink 로고 필요)
+        String sub = user.getAttribute("sub");
+        String email = user.getAttribute("email");
 
 
         return PersonalDetail.builder()
-                .providerId(id.toString())
-                .email(null)
-                .oauthProvider(OauthProvider.KAKAO)
+                .email(email)
+                .providerId(sub)
+                .oauthProvider(OauthProvider.GOOGLE)
                 .status(PersonalDetail.Status.REQUIRED_SMS)
                 .build();
     }
