+++ src/main/java/spring/fitlinkbe/support/security/provider/NaverProviderHandler.java
@@ -1,4 +1,4 @@
-package spring.fitlinkbe.infra.security.provider;
+package spring.fitlinkbe.support.security.provider;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
@@ -7,27 +7,31 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import spring.fitlinkbe.domain.common.model.PersonalDetail;
 
+import java.util.Map;
+
 import static spring.fitlinkbe.domain.common.model.PersonalDetail.OauthProvider;
 
 @Component
 @Transactional
 @RequiredArgsConstructor
-public class GoogleProviderHandler implements Oauth2ProviderHandler {
+public class NaverProviderHandler implements Oauth2ProviderHandler {
     @Override
     public boolean supports(OauthProvider provider) {
-        return provider == OauthProvider.GOOGLE;
+        return provider == OauthProvider.NAVER;
     }
 
     @Override
     public PersonalDetail handle(OAuth2UserRequest userRequest, OAuth2User user) {
-        String sub = user.getAttribute("sub");
-        String email = user.getAttribute("email");
+        Map<String, String> response = user.getAttribute("response");
+        assert response != null;
 
+        String id = response.get("id");
+        String email = response.get("email");
 
         return PersonalDetail.builder()
+                .providerId(id)
                 .email(email)
-                .providerId(sub)
-                .oauthProvider(OauthProvider.GOOGLE)
+                .oauthProvider(OauthProvider.NAVER)
                 .status(PersonalDetail.Status.REQUIRED_SMS)
                 .build();
     }
